refactor(sut): Split dumblog library to smaller libraries
This change make it so that we don't have to recompile as much when changing the
code. For example if we update only the dumblog-journal code, we don't need to
recompile metrics nor dumblog-sqlite.
