--- src/sut/dumblog/src/Dumblog/Common/Constants.hs
@@ -1,4 +1,0 @@
-module Dumblog.Common.Constants where
-
-dUMBLOG_PORT :: Int
-dUMBLOG_PORT = 8054
