+++ src/sut/dumblog/src/common/Dumblog/Common/Types.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+module Dumblog.Common.Types where
+
+import Data.Binary (Binary)
+
+newtype SeqNum = SeqNum { unSeqNum :: Int }
+  deriving newtype (Show, Binary)
