+++ src/sut/dumblog/src/common/Dumblog/Common/Utils.hs
