+++ src/sut/dumblog/src/journal/Dumblog/Journal/StateMachine.hs
